Fixing UI, so that the UI doesn't get stuck in "Loading" when you change the status of a light
There was a race condition, where when changing a Light (e.g. from off -> on), we'd emit both a change event for off, and a change event for on. This was because we emitted a "Property Changed" event in the "setProperty" method, even though the property may not actually have changed then.