Fixing ambiguous grammar for defining start/ end.
Previously we used to allow ({ start: 23:00, end: 06:00 }) as a shortcut for defining (basically) "nighttime".
However, in cases where you also want ({ start: 06:00, end: sunrise }) as a way of defining "from 06:00 -> sunrise", but when sunrise can be before 06:00, the code would go down the path of the first case, and actually treat it as 06:00 -> sunrise of tomorrow.

Now, we require the first case to be defined as ({ start: 23:00, end: '06:00 + 1d' })