Adding Smart Home Lambda, so Alexa interactions go through Karen, rather than through individual providers (e.g. Tado, Samsung, Lightwave).
This is in preparation for removing Samsung SmartThings in favour of our own controller, but also has the benefit of meaning we don't have to say "Alexa, tell Karen to ..." to do device operations via Karen (instead of the flaky 3rd party services).

We have broken the TTS Alexa as part of this, but will fix later.